django/requirements.txt
asgiref==3.5.2
distlib==0.3.4
Django==4.0.5
django-bootstrap==0.2.4
django-tastypie==0.14.4
djangorestframework==3.13.1
filelock==3.7.1
mysqlclient==2.1.0
Pillow==9.1.1
platformdirs==2.5.2
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2022.1
resources==0.0.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.14.1
django-cors-headers
django-environ